+++ DesignAutomationConsole/Program.cs
@@ -26,9 +26,6 @@ namespace DesignAutomationConsole
 
         private static async Task DA_RevitAddin_DA_Tester()
         {
-            const string RequestUri =
-                "https://github.com/ricaun-io/RevitAddin.DA.Tester/releases/latest/download/RevitAddin.DA.Tester.bundle.zip";
-
             var appName = "RevitAddin_DA_Tester";
 
             IDesignAutomationService designAutomationService = new RevitDesignAutomationService(appName)
@@ -60,6 +57,8 @@ namespace DesignAutomationConsole
 
             await designAutomationService.Initialize($".\\Bundle\\RevitAddin.DA.Tester.bundle.zip");
 
+            //const string RequestUri =
+            //    "https://github.com/ricaun-io/RevitAddin.DA.Tester/releases/latest/download/RevitAddin.DA.Tester.bundle.zip";
             //await designAutomationService.Initialize(await RequestService.Instance.GetFileAsync(RequestUri));
 
             var engineVersions = designAutomationService.CoreEngineVersions();
@@ -73,7 +72,7 @@ namespace DesignAutomationConsole
                     Input = new InputModel()
                     {
                         Text = engineVersion,
-                        Sleep = 2000
+                        Sleep = 1000
                     }
                 };
                 options.Add(option);